Enhance dataset processing and configuration options
- Updated `settings.json` to include a new `conversation_strategy` and `time_window` for improved conversation handling.
- Refactored `qa_generator.py` to initialize conversation strategies based on the new configuration options.
- Modified `strategies.py` to improve type hints and clarify strategy implementations for conversation management.
